feat: 비즈니스 로직을 hook으로 추상화하고, 결합도를 낮춤
